+++ saleor/app/models.py
@@ -58,6 +58,7 @@ class App(ModelWithMetadata):
     manifest_url = models.URLField(blank=True, null=True)
     version = models.CharField(max_length=60, blank=True, null=True)
     audience = models.CharField(blank=True, null=True, max_length=256)
+    is_installed = models.BooleanField(default=True)
     objects = AppManager()
 
     class Meta(ModelWithMetadata.Meta):