+++ saleor/app/tests/test_app_tasks.py
@@ -16,6 +16,7 @@ def test_install_app_task(app_installation):
     app = App.objects.filter(name=app_installation.app_name).first()
     assert app
     assert app.is_active is False
+    assert app.is_installed
 
 
 @pytest.mark.vcr